% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/annotateEvents.R
\name{annotateEvents}
\alias{annotateEvents}
\title{Will match eigensentences to events and given unique labels to words,
sentences etc.}
\usage{
annotateEvents(sentencesIn, words, eventtimes, eventnames)
}
\arguments{
\item{sentencesIn}{the list of possible sentences}

\item{words}{the list of possible words}

\item{eventtimes}{the index to the temporal volumes at which events occur}

\item{eventnames}{a name for every event so of equal length to eventtimes}
}
\value{
returns a data frame
}
\description{
Returns a dataframe for every event with annotated event times, word
categories (V, N,... ), sentence descriptors (labels, other variables) etc.
}
\examples{
data(reuters_words,package="RKRNS")
  data(sentences,package="RKRNS")
  eventtimes<-c(10,25,99)
  eventnames<-list(sentences$Sentence[1],sentences$Sentence[55],sentences$Sentence[19])
  eventdf<-annotateEvents( sentences$Sentence, reuters_words$WhichWord, eventtimes, unlist(eventnames) )
}
\author{
Avants BB
}

