\name{RKRNS-package}
\alias{RKRNS-package}
\alias{RKRNS}
\docType{package}
\title{
Research on decoding sentences from BOLD data
}
\description{
Filters fMRI in order to identify signal that may relate to sentence
reading.  Features spaces based on spectral representations of the
data.  Also seeks to help understand classification patterns.
}
\details{
\tabular{ll}{
Package: \tab RKRNS\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-07-05\cr
License: \tab GPL-2\cr
}
Significant content in the demo folder of the source package.  Assumes the data is stored elsewhere.  The vignette also provides an overview of the ideas here.
}
\author{
Brian B. Avants

Maintainer: Brian B. Avants <stnava@gmail.com>
}
\references{
See the vignette
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[https://github.com/stnava/ANTsR]{ANTsR}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
