% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rkrnsRanker.R
\name{rkrnsRanker}
\alias{rkrnsRanker}
\title{use random forest or svm to predict and rank sentences from features}
\usage{
rkrnsRanker(mydfin, trainingRows, targetSuccessRate = 0.15,
  whichmodel = "randomForest", verbose = FALSE)
}
\description{
takes a dataframe that contains a column named lab such that mydf$lab gives
the true labeling of the data. trainingRows is a vector identifying which
rows should be training data.  success is defined by a prediction
probability being within the top targetSuccessRate*100 percent.  Returns
various success metrics.
}
\examples{
mylabs<-rep(c("a","b","c"),5)
voxeldata<-replicate(100, rnorm(length(mylabs)))
mydf<-data.frame( lab=mylabs, vox=voxeldata)
rfr<-rkrnsRanker( mydf, 1:round(nrow(mydf)/2), whichmodel='svm' )
print( rfr$successPercent )
}

