% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sccanBasedDecoder.R
\name{sccanBasedDecoder}
\alias{sccanBasedDecoder}
\title{Simple eigenSentence estimation function.}
\usage{
sccanBasedDecoder(eventdata, designmat, boldFeatureMatrix, sentenceSpace,
  mysparse = c(-0.1, -0.1), nvecs = 5, its = 1, smooth = 0,
  cthresh = 0, mask = NA, strategy = NA, doEanat = F, joinEanat = F,
  outputfileprefix = "sccanBasedDecoder", interleave = FALSE,
  sentenceTransformation = "none", trainset = NA, locwordlist = "")
}
\arguments{
\item{eventdata}{output from annotateEvents}

\item{designmat}{sentences part of the assembled design matrix}

\item{boldFeatureMatrix}{a feature matrix for bold data, currently
spatiotemporal}

\item{sentenceSpace}{the sentence feature space}

\item{mask}{a 4D mask}

\item{doEanat}{also run a sparse PCA on the data}

\item{otherparameters}{see sparseDecom2 documentation}
}
\value{
a data frame with annotation and ground truth vs prediction
annotated ...
}
\description{
Applies a function to a matrix representation of a sentence to get an
eigensentence map.
}
\examples{
\dontrun{
 ccaresults<-sccanBasedDecoder( eventdata, dmats, ccafeatspace , sentspace )
}
}
\author{
Avants BB
}

