% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/showMemoryUse.R
\name{showMemoryUse}
\alias{showMemoryUse}
\title{gets all obejcts with their sizes shamelessly copied from
http://stackoverflow.com/questions/1358003/tricks-to-manage-the-available-memory-in-an-r-session}
\usage{
showMemoryUse(sort = "size", decreasing = TRUE, limit = 5)
}
\description{
gets all obejcts with their sizes shamelessly copied from
http://stackoverflow.com/questions/1358003/tricks-to-manage-the-available-memory-in-an-r-session
}

