% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bold2betasCCA.R
\name{bold2betasCCA}
\alias{bold2betasCCA}
\title{Convert a bold time series and design matrix to event-wise cca-based betas}
\usage{
bold2betasCCA(boldmatrix, designmatrix, blockNumb, bl = 12, baseshift = 5,
  mask = NA, sparseness = c(0, 0), multievents = FALSE, polydegree = 10,
  bestvoxnum = 50, nvecs = 5, whichcols = NA, mycoption = 1, its = 10,
  onlyhrf = FALSE)
}
\arguments{
\item{boldmatrix}{input raw bold data in time by space matrix}

\item{designmatrix}{input design matrix - binary/impulse entries for event
related design, blocks otherwise}

\item{blockNumb}{numbers for the rows that should be treated together as
runs}

\item{bl}{basis length for hrf estimation}

\item{baseshift}{number of time points to ignore post-event onset}

\item{polydegree}{number of polynomial predictors}

\item{nvecs}{number of cca predictors to explore e.g. 5}
}
\value{
returns a list with relevant output
}
\description{
Uses cca and nuisance variables to estimate multivariate betas per event.
}
\examples{

# get example image
fn<-paste(path.package("RKRNS"),"/extdata/111157_mocoref_masked.nii.gz",sep="") 
eximg<-antsImageRead(fn,3)
fn<-paste(path.package("RKRNS"),"/extdata/subaal.nii.gz",sep="") 
mask<-antsImageRead(fn,3)
bb<-simulateBOLD(option="henson",eximg=eximg,mask=mask)
mat<-timeseries2matrix( bb$simbold, bb$mask )
runs<-bb$desmat$Run; 
### cca betas
mysp<-c( -0.01, -0.9 )
btsc<-bold2betasCCA( data.matrix(mat)  , bb$desmat[,1:4], blockNumb=runs,
     bl=25, baseshift=0, sparseness=mysp, bestvoxnum=10, mask=mask, 
     polydegree=1, mycoption=1, its=12, onlyhrf=FALSE )
plot(ts(t(btsc$runhrfs)))
# from here see the help for bold2betas

}
\author{
Avants BB
}
