% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenSentences.R
\name{eigenSentences}
\alias{eigenSentences}
\title{Simple eigenSentence estimation function.}
\usage{
eigenSentences(wordembed, functiontoapply = sum, normalize = F,
  eigsentbasislength = NA, sentencesIn = NA, eventdata = NA)
}
\arguments{
\item{wordembed}{the words and their vector embedding - a data frame with n
basis length columns}

\item{functiontoapply}{e.g. mean, max, min, a custom function --- needs to
work with apply}

\item{normalize}{normalize the magnitude of the eigsentences to be 1}

\item{eigsentbasislength}{less than or equal to the full basis length (100)}
}
\value{
eigensentence matrix is output
}
\description{
Applies a function to a matrix representation of a sentence to get an
eigensentence map.
}
\examples{

  data(reuters_words,package="RKRNS")
  esent<-eigenSentences( reuters_words )

}
\author{
Avants BB
}
