% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterfMRI4KRNS.R
\name{filterfMRI4KRNS}
\alias{filterfMRI4KRNS}
\title{Filtering options specific to KRNS}
\usage{
filterfMRI4KRNS(imat, tr = as.numeric(0.5), filterlowfrequency = 0.1,
  filterhighfrequency = 1, trendfrequency = 3, trendfrequency2 = NA,
  winsorval = 0.01, removeSentLengthEffects = NA, removeEventOverlap = NA)
}
\arguments{
\item{filterlowfrequency}{get rid of variation that takes many seconds}

\item{filterhighfrequency}{get rid of variation that takes few seconds}

\item{trendfrequency}{time series trend option}

\item{winsorval}{remove spikes in signal given this winsorization value}

\item{removeSentLengthEffects}{nuisance vector of size nrow of matrix}

\item{removeEventOverlap}{nuisance vector of size nrow of matrix}

\item{mat}{input bold time series matrix}
}
\value{
matrix is output
}
\description{
Filters the input matrix using butterworth filtering and time-series trends.
}
\examples{

fmat<-filterfMRI4KRNS( mat ) 

}
\author{
Avants BB
}
