% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/finiteImpulseResponseDesignMatrix.R
\name{finiteImpulseResponseDesignMatrix}
\alias{finiteImpulseResponseDesignMatrix}
\title{Simple finiteImpulseResponseDesignMatrix function.}
\usage{
finiteImpulseResponseDesignMatrix(x, n = 1, baseshift = 0)
}
\arguments{
\item{n}{shift amount}

\item{baseshift}{baseline shift amount so total shift is n+baseshift}

\item{mat}{input matrix}
}
\value{
matrix is output
}
\description{
Convert a design matrix (or any matrix) to a time-shifted version of itself
as needed by finite impulse response models used when deconvolving
hemodynamic response functions.
}
\examples{
mat<-diag(10)
wmat<-finiteImpulseResponseDesignMatrix( mat , 5 )
}
\author{
Avants BB
}

