% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spatioTemporalProjectionImage.R
\name{spatioTemporalProjectionImage}
\alias{spatioTemporalProjectionImage}
\title{Projects a 4D image to a 3D and 1D image (ND to 1 and D minus 1)}
\usage{
spatioTemporalProjectionImage(spacetimematrixIn, summaryFUN = mean,
  nDminus1ref = NA)
}
\arguments{
\item{summaryFUN}{how to summarize the spatial part across time}

\item{nDminus1ref}{reference spatial image}

\item{spacetimevec}{input space time matrix or antsImage of length equal to
nvoxels in a mask times the number of time points}
}
\value{
list(spacefunction=myestimatedbrainregionsval,
timefunction=myestimatedhrf, spaceimage=spaceimage )
}
\description{
Applies a summary function to a spatiotemporal matrix to get lower
dimensional representations.
}
\examples{
\dontrun{
img<-antsImageRead("TEST0View1vec000.nii.gz",4)
ff<-spatioTemporalProjectionImage( img, sum, mask )
antsImageWrite(ff$spaceimage,'temp.nii.gz')
plot(ff$timefunction,type='l')
}
}
\author{
Avants BB
}

